Refactor database and model configs into one file
